# goblas: fix Ddot for unaligned slices