No real change in the benchmark results
(Go 1.4 on Intel(R) Xeon(R) CPU E5-2660 0 @ 2.20GHz):
```
benchmark                          old ns/op     new ns/op     delta
BenchmarkDdotSmallBothUnitary      15.3          15.3          +0.00%
BenchmarkDdotSmallIncUni           21.5          21.5          +0.00%
BenchmarkDdotSmallUniInc           22.0          22.0          +0.00%
BenchmarkDdotSmallBothInc          21.5          21.5          +0.00%
BenchmarkDdotMediumBothUnitary     505           505           +0.00%
BenchmarkDdotMediumIncUni          1186          1188          +0.17%
BenchmarkDdotMediumUniInc          1178          1178          +0.00%
BenchmarkDdotMediumBothInc         1220          1213          -0.57%
BenchmarkDdotLargeBothUnitary      59322         59327         +0.01%
BenchmarkDdotLargeIncUni           163780        163693        -0.05%
BenchmarkDdotLargeUniInc           141651        141706        +0.04%
BenchmarkDdotLargeBothInc          203541        203794        +0.12%
BenchmarkDdotHugeBothUnitary       9586039       9586340       +0.00%
BenchmarkDdotHugeIncUni            31228460      31128999      -0.32%
BenchmarkDdotHugeUniInc            21437424      21476630      +0.18%
BenchmarkDdotHugeBothInc           40115032      40122703      +0.02%
```
Fixes #77