@btracey 2015-01-04
LGTM

Would you mean explaining a bit for my education? I tried googling but the discussions were pretty technical. I can see the affect of the change, and understand the differences. However, it seems like there are two instructions that are equally fast, where one only works on aligned data, and the other works on any data. Why would one want to use MOVAPD when you could use MOVUPD instead? Is there a theoretic speed penalty? Is it just that you didn't understand? I'm not trying to be accusatory, just trying to keep up on top of code development. Thanks for the changes.