@btracey 2015-01-04
Just a note, early merge because it fixes ddot.We should still consider the incremented test if appropropriate.