@fhs 2015-01-04
MODAPD is suppose to be faster than MOVUPD, but in practice, at least on the CPU I ran the benchmarks on, it doesn't make much difference. It general alignment is a good thing. See: https://groups.google.com/d/msg/golang-nuts/HFjwPFYrCqg/oSttxY8ajr4J