@btracey 2015-01-04
I'm going to leave a related note on that thread for future readers. Feel free to chime in if you have things to add.

Thanks for the fix.